some refactoring for behavior search
