+++ src/jabs/behavior_search/__init__.py
@@ -1,0 +1,9 @@
+"""This package provides the non-UI classes and functions for behavior search."""
+
+from .behavior_search_util import (
+    BehaviorSearchQuery,
+    LabelBehaviorSearchQuery,
+    PredictionLabelSearchQuery,
+    SearchHit,
+    search_behaviors,
+)
